Changes so that entire scenario can be inserted into DB from XML file.

+++ src/main/java/edu/berkeley/path/model_objects/scenario/Sensor.java
@@ -268,7 +268,7 @@ public class Sensor extends edu.berkeley.path.model_objects.jaxb.Sensor {
 	 */
 	@Override
 	public Integer getLaneNumber() {
-		return super.getLaneNumber().intValue();
+		return super.getLaneNumber();
 	}
